fix: stabilize i18n initialization and move ai utils
