+++ utils/aiUtils.ts
